+++ applications/fishing-map/src/features/map/popups/PopupWrapper.tsx
@@ -1,4 +1,5 @@
 import React, { Fragment } from 'react'
+import { useSelector } from 'react-redux'
 import cx from 'classnames'
 import { groupBy } from 'lodash'
 import type { Anchor } from '@globalfishingwatch/mapbox-gl'
@@ -7,7 +8,8 @@ import { Generators } from '@globalfishingwatch/layer-composer'
 import { DataviewCategory } from '@globalfishingwatch/api-types/dist'
 import { TooltipEvent } from 'features/map/map.hooks'
 import { POPUP_CATEGORY_ORDER } from 'data/config'
-import { PRESENCE_POC_INTERACTION, USE_PRESENCE_POC } from 'features/datasets/datasets.slice'
+import { PRESENCE_POC_INTERACTION } from 'features/datasets/datasets.slice'
+import { selectDebugOptions } from 'features/debug/debug.slice'
 import styles from './Popup.module.css'
 import FishingTooltipRow from './FishingLayers'
 import PresenceTooltipRow from './PresenceLayers'
@@ -36,6 +38,8 @@ function PopupWrapper({
   onClose,
   anchor,
 }: PopupWrapperProps) {
+  const debugOptions = useSelector(selectDebugOptions)
+
   if (!event) return null
 
   const visibleFeatures = event.features.filter((feature) => feature.visible)
@@ -72,7 +76,7 @@ function PopupWrapper({
               return features.map((feature, i) => {
                 if (
                   feature.temporalgrid?.sublayerInteractionType === 'presence-detail' ||
-                  (USE_PRESENCE_POC &&
+                  (debugOptions.presenceTrackPOC &&
                     feature.temporalgrid?.sublayerInteractionType === PRESENCE_POC_INTERACTION)
                 ) {
                   return (