+++ applications/fishing-map/src/features/search/Search.tsx
@@ -356,6 +356,7 @@ function Search() {
                         geartype,
                         origin,
                         casco,
+                        nationalId,
                         matricula,
                         dataset,
                         firstTransmissionDate,
@@ -413,6 +414,12 @@ function Search() {
                                 <div className={styles.property}>
                                   <label>{t('vessel.matricula', 'Matricula')}</label>
                                   <span>{matricula}</span>
+                                </div>
+                              )}
+                              {nationalId && (
+                                <div className={styles.property}>
+                                  <label>{t('vessel.nationalId', 'National Id')}</label>
+                                  <span>{nationalId}</span>
                                 </div>
                               )}
                               {casco && (