+++ applications/vessel-history/src/utils/flags.ts
@@ -1,4 +1,5 @@
 import flags from 'data/flags'
+import i18n, { t } from 'features/i18n/i18n'
 
 type Flag = { id: string; label: string }
 export const getFlagById = (id: string): Flag | undefined => {
@@ -9,3 +10,13 @@ export const getFlagById = (id: string): Flag | undefined => {
     label: flag.label,
   }
 }
+
+export const getFlags = (lng = i18n.language): Flag[] =>
+  flags
+    .map((flag) => {
+      return {
+        ...flag,
+        label: t(`flags:${flag.id}`, { lng }) || flag.label,
+      }
+    })
+    .sort((a, b) => a.label.localeCompare(b.label))