+++ packages/api-types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @globalfishingwatch/api-types
 
+## 5.3.1
+
+### Patch Changes
+
+- 998b9274: add nationalId in vessel type definition
+
 ## 5.3.0
 
 ### Minor Changes